Revert "WIP fix bug on null data"
This reverts commit 698a2f65ca82c054c6623bbe2888bfbb1bca2b46.

+++ Vaccine_Analysis_by_Country/Main Project/MongoDb_Transformations.py
@@ -1,5 +1,4 @@
 # Importing necessary packages to allow to transform data in MongoDB.
-from xml.dom.minidom import Document
 import pymongo.errors
 from pymongo import MongoClient
 from MongoDB_Import import db
@@ -42,90 +41,57 @@ print(list(db.Countries.find({'Country': "Poland"})), '\n')
 
 # Creating collection which will be a combination of all three datasets (df_1, df_2 and df_3).
 try:
+    if db.Total_Vaccinations_Worldwide.count() > 0:
+        db.Total_Vaccinations_Worldwide.drop()
 
-    db.Total_Vaccinations_Worldwide.drop()
     db.create_collection('Total_Vaccinations_Worldwide')
 
 except pymongo.errors.PyMongoError as pyerror2:
     print('Oh no an error: ', pyerror2)
-mongoDbQuery = [{
-    "$match": {
-        "$and": [
-            {"date": {
-                '$gte': '2021-01-01',
-                '$lte': '2021-11-24'
-            }},
-
-            {"total_vaccinations": {
-                "$ne": 'null'}},
-
-            {"people_fully_vaccinated": {
-                "$ne": 'null'}}
-        ]
-    }
-},
-    {"$group": {
-        "_id": "$location",
-        "Iso_Code": {"$last": "$iso_code"},
-        "Date_of_Last_Record": {"$last": "$date"},
-        "Total_Vaccinations_Administered": {"$last": "$total_vaccinations"},
-        "Total_Boosters_Administered": {"$last": "$total_boosters"},
-        "Number_of_People_Fully_Vaccinated": {"$last": "$people_fully_vaccinated"},
-    }}]
 
 # Populating the Total_Vaccinations_Worldwide
 try:
     # A for loop to populate Total_Vaccinations_Worldwide collection using the data from df_1 collection.
-    total_vac_data = db.df_1.aggregate(mongoDbQuery)
+    total_vac_data = db.df_1.aggregate([
+        {
+           "$match": {
+            "$and":[
+                    {"date": { "$gte":'2021-01-01',
+                               "$lte":'2021-11-24'}},
+
+                    {"total_vaccinations":{
+                                "$exists": "true",
+                                "$ne": "null" }},
+
+                    {"people_fully_vaccinated":{
+                                "$exists": "true",
+                                "$ne": "null" }}
+            ]
+        }
+    },
+        {"$group": {
+            "_id": "$location",
+            "Iso_Code": {"$last": "$iso_code"},
+            "Total_Vaccinations_Administered": {"$last": "$total_vaccinations"},
+            "Total_Boosters_Administered": {"$last": "$total_boosters"},
+            "Number_of_People_Fully_Vaccinated": {"$last": "$people_fully_vaccinated"},
+        }}])
 
     for data in total_vac_data:
-        if data['Total_Vaccinations_Administered'] is None or data['Number_of_People_Fully_Vaccinated'] is None:
-            country = data['_id']
-
-            newDataForCountry = db.df_1.aggregate([{
-                "$match": {
-                    "$and": [
-                        {"date": {
-                            '$gte': '2021-01-01',
-                            '$lte': '2021-11-24'
-                        }},
-
-                        {"total_vaccinations": {
-                            "$ne": 'null'}},
-
-                        {"people_fully_vaccinated": {
-                            "$ne": 'null'}},
-
-                        {"location": country}
-                    ]
-                }
-            },
-                {"$group": {
-                    "_id": "$location",
-                    "Date_of_Last_Record": {"$last": "$date"},
-                    "Total_Vaccinations_Administered": {"$last": "$total_vaccinations"},
-                    "Number_of_People_Fully_Vaccinated": {"$last": "$people_fully_vaccinated"},
-                }}])
-
-            for newData in newDataForCountry:
-                data['_id'] = data['_id']
-                data['Iso_Code'] = data['Iso_Code']
-                data['Date_of_Last_Record'] = newData['Date_of_Last_Record']
-                data['Total_Vaccinations_Administered'] = newData['Total_Vaccinations_Administered']
-                data['Number_of_People_Fully_Vaccinated'] = newData['Number_of_People_Fully_Vaccinated']
-
         db.Total_Vaccinations_Worldwide.insert_one(
             {
                 'Country': data['_id'],
                 'ISO_Code': data['Iso_Code'],
-                'Date_of_Last_Record': data['Date_of_Last_Record'],
                 'Total_Vaccinations_Administered': data['Total_Vaccinations_Administered'],
                 'Total_Boosters_Administered': data['Total_Boosters_Administered'],
                 'Number_of_People_Fully_Vaccinated': data['Number_of_People_Fully_Vaccinated']
             }
         )
+
 except pymongo.errors.PyMongoError as error:
     print('There was an error populating the Countries table:', error)
 
+
 # Check if collection exists
 print('Here are our collections:', '\n', my_db.list_collection_names(), '\n')
+
